+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - `cli`: Add support to show description for env and desk
+- `cli`: Add support for alias commands
 
 ### Changed